requirements/requirements_main.txt
pdfplumber
pandas
langchain
langchain-core
langchain_huggingface
langchain-anthropic
langchain-openai
langchain-community
python-dotenv>=1.1.1
notebook
transformers
datasets>=2.2.0
pyarrow
numpy
requests>=2.26.0
sentence_transformers>=2.2.0
torch
tqdm
rich
InstructorEmbedding
faiss-cpu
pyperclip
ipywidgets

# qdrant specific
qdrant-client
langchain-qdrant
docling
docling-core
easyocr
fastembed

# evaluator
openevals

# observalbility and evaluate
mlflow

# fast api
fastapi[all]
uvicorn 

# Checkpointer progrestsql
psycopg[binary]
langgraph-checkpoint-postgres 

modal

# fast api utils
slowapi